requirements.txt
python-telegram-bot
beautifulsoup4
bs4
